+++ frontend/app/(dashboard)/role/page.jsx
@@ -1,5 +1,5 @@
 "use client";
-import SheetDrawer from "@/components/ui-main/Sheet";
+
 import ReusableTable from "@/components/ui-main/Table";
 import { Button } from "@/components/ui/button";
 import {
@@ -10,9 +10,15 @@ import { PlusIcon } from "lucide-react";
 import { useRouter } from "next/navigation";
 import React, { useState } from "react";
 import { toast } from "react-toastify";
-import AddRoleForm from "./_components/add-role";
+
+import usePermission from "@/hooks/useCheckPermission";
 
 function page() {
+  const canAddRole = usePermission("ADD_ROLE");
+  const canListRole = usePermission("LIST_ROLE");
+  const canEditRole = usePermission("UPDATE_ROLE");
+  const canDeleteRole = usePermission("DELETE_ROLE");
+
   const [searchTerm, setSearchTerm] = useState("");
   const router = useRouter();
   const { data: roleList, isLoading } = useGetRoleListQuery({
@@ -58,30 +64,36 @@ function page() {
       Header: "Description",
       accessor: "description",
     },
+  ];
 
-    {
+  if (canEditRole || canDeleteRole) {
+    columns.push({
       Header: "Actions",
       accessor: "actions",
       render: (value, row) => (
         <div className="flex gap-2">
-          <Button
-            className="bg-blue-500 text-white px-2 py-1 rounded text-sm hover:bg-blue-600"
-            onClick={() =>
-              router.push(`role/add-role?id=${row?._id}&mode=edit`)
-            }
-          >
-            Edit
-          </Button>
-          <Button
-            className="bg-red-500 text-white px-2 py-1 rounded text-sm hover:bg-red-600"
-            onClick={() => handleDelete(row?._id)}
-          >
-            Delete
-          </Button>
+          {canEditRole && (
+            <Button
+              className="bg-blue-500 text-white px-2 py-1 rounded text-sm hover:bg-blue-600"
+              onClick={() =>
+                router.push(`role/add-role?id=${row?._id}&mode=edit`)
+              }
+            >
+              Edit
+            </Button>
+          )}
+          {canDeleteRole && (
+            <Button
+              className="bg-red-500 text-white px-2 py-1 rounded text-sm hover:bg-red-600"
+              onClick={() => handleDelete(row?._id)}
+            >
+              Delete
+            </Button>
+          )}
         </div>
       ),
-    },
-  ];
+    });
+  }
 
   return (
     <div className="p-6 bg-gray-100">
@@ -89,22 +101,33 @@ function page() {
         <div>
           <h1 className="text-2xl font-bold mb-6">Role Management</h1>
         </div>
-        <div>
-          <Button icon={PlusIcon} onClick={() => router.push("role/add-role")}>
-            Add Role
-          </Button>
-        </div>
+        {canAddRole && (
+          <div>
+            <Button
+              icon={PlusIcon}
+              onClick={() => router.push("role/add-role")}
+            >
+              Add Role
+            </Button>
+          </div>
+        )}
       </div>
 
-      <ReusableTable
-        columns={columns}
-        data={roleData}
-        loading={isLoading}
-        emptyMessage={roleList?.message || "No role found"}
-        searchValue={searchTerm}
-        searchBarPlaceholder="Search by name"
-        onSearchChange={(val) => setSearchTerm(val)}
-      />
+      {canListRole ? (
+        <ReusableTable
+          columns={columns}
+          data={roleData}
+          loading={isLoading}
+          emptyMessage={roleList?.message || "No role found"}
+          searchValue={searchTerm}
+          searchBarPlaceholder="Search by name"
+          onSearchChange={(val) => setSearchTerm(val)}
+        />
+      ) : (
+        <section>
+          <div>You don't have permission to view this</div>
+        </section>
+      )}
     </div>
   );
 }
